Minor refactor and add comment
