+++ app/admin/__init__.py
@@ -1,5 +1,8 @@
+# The routes are all for admins.
+
 from flask import Blueprint
 
 admin_bp = Blueprint("admin", __name__)
 
 from . import routes
+from . import backend_routes
