+++ app/admin/routes.py
@@ -1,24 +1,19 @@
-import csv
 from flask import request, render_template, current_app, make_response
 from flask.helpers import flash
 from flask_login import login_required
-from ..forms import ReportsFilterForm, AddOneUserForm, AddUsersByFileForm
 from . import admin_bp
+from .helper import csv_handler
+from ..forms import ReportsFilterForm, AddOneUserForm, AddUsersByFileForm
 from ..database.db_helper import (
     render_system_setting,
-    delete,
-    update,
-    insert,
     render_records,
     render_users,
     add_users,
-    del_users,
-    update_users,
 )
 from ..users import admin_required
-from app import admin
 
 
+# The page allows admins to browse all reports and make response to the reports.
 @admin_bp.route("/admin_dashboard/", methods=["GET", "POST"])
 @admin_bp.route("/admin_dashboard/<int:page>", methods=["GET", "POST"])
 @admin_required
@@ -29,15 +24,15 @@ def dashboard_page(page=1):
     if request.method == "GET":
         current_app.logger.info("GET /admin_dashboard")
         Filter = {}
-        if username:= request.cookies.get("username"):
+        if username := request.cookies.get("username"):
             Filter["username"] = username
-        if classnum:= request.cookies.get("classnum"):
+        if classnum := request.cookies.get("classnum"):
             Filter["classnum"] = classnum
         return render_template(
             "admin_dashboard.html",
             records=render_records(Filter=Filter, page=page),
             form=form,
-            statuses=render_system_setting()[3]
+            statuses=render_system_setting()[3],
         )
     if request.method == "POST":
         if form.validate_on_submit():
@@ -53,7 +48,10 @@ def dashboard_page(page=1):
 
             response = make_response(
                 render_template(
-                    "admin_dashboard.html", records=render_records(Filter), form=form, statuses=render_system_setting()[3]
+                    "admin_dashboard.html",
+                    records=render_records(Filter),
+                    form=form,
+                    statuses=render_system_setting()[3],
                 )
             )
             response.delete_cookie("username")
@@ -65,20 +63,7 @@ def dashboard_page(page=1):
             current_app.logger.info("POST /admin_dashboard: Invalid submit")
 
 
-@admin_bp.route("/admin_dashboard_backend", methods=["POST", "DELETE"])
-@admin_required
-@login_required
-def admin_dashboard_backend_page():
-    if request.method == "POST":
-        data = request.get_json(force=True)
-        print(data)
-        return "OK"
-    if request.method == "DELETE":
-        data = request.get_json(force=True)
-        print(data)
-        return "OK"
-
-
+# The page allows admins to modify system setting. For example, they can add more buildings, offices and so on to the system.
 @admin_bp.route("/system", methods=["GET"])
 @admin_required
 @login_required
@@ -95,48 +80,7 @@ def system_page():
         )
 
 
-@admin_bp.route("/system_modification", methods=["POST", "DELETE", "UPDATE"])
-@admin_required
-@login_required
-def system_modification_page():
-    if request.method == "POST":
-        # Add
-        current_app.logger.info("POST /system_modification")
-        data = request.get_json(force=True)
-        if data.get("office") != None:
-            args = {"description": data["value"], "office_id": data["office"]}
-        else:
-            args = {"description": data["value"]}
-        if insert(data["category"], args):
-            return "OK"
-        else:
-            return "Error", 400
-    if request.method == "DELETE":
-        # Delete
-        current_app.logger.info("DELETE /system_modification")
-        data = request.get_json(force=True)
-        if delete(data["category"], data["id"]):
-            return "OK"
-        else:
-            return "Error", 400
-    if request.method == "UPDATE":
-        # Update
-        current_app.logger.info("UPDATE /system_modification")
-        data = request.get_json(force=True)
-        category = data[0]["category"]
-        for r in data[1:]:
-            if not update(
-                category,
-                {
-                    "id": r["id"],
-                    "description": r["description"],
-                    "sequence": r["sequence"],
-                },
-            ):
-                return "Error", 400
-        return "OK"
-
-
+# The page allows admins to add, edit and delete users.
 @admin_bp.route("/manage_user/", methods=["GET", "POST"])
 @admin_bp.route("/manage_user/<int:page>", methods=["GET", "POST"])
 @admin_required
@@ -171,48 +115,19 @@ def manage_user_page(page=1):
                 flash(", ".join(already_exists) + " 已經存在", category="alert")
         else:
             current_app.logger.info("POST /manage_user: Invalid submit")
+
         # Add users by csv
         if form_csv.validate_on_submit():
             current_app.logger.info("POST /manage_user")
             csv_file = form_csv.csv_file.data
-            rawdata = csv_file.read()
             # data format: [{"username": "zxc", "name": "zxc", "password": "123", "classnum": "1400"}]
-            try:
-                data = [
-                    row for row in csv.DictReader(rawdata.decode("utf-8").splitlines())
-                ]
-            except UnicodeDecodeError:
-                data = [
-                    row for row in csv.DictReader(rawdata.decode("big5").splitlines())
-                ]
-            except Exception:
-                data = [
-                    row for row in csv.DictReader(rawdata.decode("ANSI").splitlines())
-                ]
-            except:
-                return "Error"
-            if already_exists := add_users(data):
-                flash(", ".join(already_exists) + " 已經存在", category="alert")
+            if not (data := csv_handler(csv_file.read())):
+                flash("Bad encoding.", category="alert")
+            else:
+                if already_exists := add_users(data):
+                    flash(", ".join(already_exists) + " 已經存在", category="alert")
         else:
             current_app.logger.info("POST /manage_user: Invalid submit")
         return render_template(
             "manage_user.html", form=form, form_csv=form_csv, users=render_users()
         )
-
-
-@admin_bp.route("/manage_user_backend", methods=["DELETE", "UPDATE"])
-@admin_required
-@login_required
-def manage_user_backend_page():
-    if request.method == "DELETE":
-        # Delete user
-        current_app.logger.info("DELETE /manage_user_backend")
-        data = request.get_json(force=True)
-        del_users([data["user_id"]])
-        return "OK"
-    if request.method == "UPDATE":
-        # Update user
-        current_app.logger.info("UPDATE /manage_user_backend")
-        data = request.get_json(force=True)
-        update_users([data])
-        return "OK"
