+++ app/user/__init__.py
@@ -1,3 +1,5 @@
+# The routes are for students and admins as well.
+
 from flask import Blueprint
 
 user_bp = Blueprint("user", __name__)
