+++ app/user/routes.py
@@ -13,6 +13,7 @@ from ..mail_helper import send_report_mail
 from . import user_bp
 
 
+# The page allows students to report broken items.
 @user_bp.route("/report", methods=["GET", "POST"])
 @login_required
 def report_page():
@@ -44,6 +45,7 @@ def report_page():
             return render_template("report.html", form=form)
 
 
+# The page is student's dashboard, and he or she can browse all his or her reports here.
 @user_bp.route("/dashboard/", methods=["GET"])
 @user_bp.route("/dashboard/<int:page>", methods=["GET"])
 @login_required
@@ -57,6 +59,7 @@ def dashboard_page(page=1):
     )
 
 
+# The page allows users, whether they are students or admins, to set their emails and passwords.
 @user_bp.route("/user_setting", methods=["GET", "POST"])
 @login_required
 def user_setting_page():
