+++ templates/admin_dashboard.html
@@ -24,7 +24,7 @@
         }
     </script>
     <link rel="icon" href="{{ url_for('static', filename='HSNU.png') }}">
-    <title>Admin Dashboard</title>
+    <title>控制台</title>
     {% if records["pages"] != 0 %}
     {% if records["page"] < 1 %} <script>window.location.replace("1");</script>{% endif %}
         {% if records["page"] > records["pages"] %}
@@ -43,7 +43,7 @@
             <span class="right"><a href="/logout">登出</a></span>
             <span class="right"><a href="/user_setting">使用者設定</a></span>
         </nav>
-        <h1>Admin Dashboard</h1>
+        <h1>控制台</h1>
         <div id="sidebar">
             <h2>篩選器</h2>
             <form action="/admin_dashboard/" method="POST">
