+++ templates/manage_user.html
@@ -95,13 +95,13 @@
         <div id="main_content">
             <div id="messages">{{ display_msg(messages) }}</div>
             <div class="pages">
-                <a href="1" {% if records['page'] <=1 %} class="end" {% endif %}>&#8249;&#8249;</a>
-                <a href="{{ records['page'] - 1 }}" {% if records['page'] <=1 %} class="end" {% endif %}>&#8249;</a>
-                <span>{{ records["page"] }} / {{ records["pages"] }}</span>
-                <a href="{{ records['page'] + 1 }}" {% if records['page']>= records['pages'] %}
+                <a href="1" {% if users['page'] <=1 %} class="end" {% endif %}>&#8249;&#8249;</a>
+                <a href="{{ users['page'] - 1 }}" {% if users['page'] <=1 %} class="end" {% endif %}>&#8249;</a>
+                <span>{{ users["page"] }} / {{ users["pages"] }}</span>
+                <a href="{{ users['page'] + 1 }}" {% if users['page']>= users['pages'] %}
                     class="end" {% endif
                     %}>&#8250;</a>
-                <a href="{{ records['pages'] }}" {% if records['page']>= records['pages'] %}
+                <a href="{{ users['pages'] }}" {% if users['page']>= users['pages'] %}
                     class="end" {% endif
                     %}>&#8250;&#8250;</a>
             </div>
@@ -120,18 +120,18 @@
                 <div>
                     <span id="{{ user['id'] }}">{{ user["username"] }}</span>
                     {% if user["is_marked_deleted"] %}
-                    <input value="{{ user[" name"] }}" type="text" disabled></input>
-                    <input value="{{ user[" classnum"] }}" type="text" disabled></input>
+                    <input value="{{ user["name"] }}" type="text" disabled></input>
+                    <input value="{{ user["classnum"] }}" type="text" disabled></input>
                     <input type="password" placeholder="未更改" disabled>
-                    <input value="{{ user[" email"] }}" class="email" type="text" disabled></input>
+                    <input value="{{ user["email"] }}" class="email" type="text" disabled></input>
                     <input type="checkbox" {% if user["is_admin"] %} checked {% endif %} disabled></input>
                     <input type="checkbox" {% if user["is_valid"] %} checked {% endif %} disabled></input>
                     <input type="checkbox" checked disabled></input>
                     {% else %}
-                    <input value="{{ user[" name"] }}" type="text"></input>
-                    <input value="{{ user[" classnum"] }}" type="text"></input>
+                    <input value="{{ user["name"] }}" type="text"></input>
+                    <input value="{{ user["classnum"] }}" type="text"></input>
                     <input type="password" placeholder="未更改">
-                    <input value="{{ user[" email"] }}" class="email" type="text"></input>
+                    <input value="{{ user["email"] }}" class="email" type="text"></input>
                     <input type="checkbox" {% if user["is_admin"] %} checked {% endif %}></input>
                     <input type="checkbox" {% if user["is_valid"] %} checked {% endif %}></input>
                     <input type="checkbox" disabled></input>
@@ -142,15 +142,13 @@
                 {% endfor %}
             </div>
             <div class="pages">
-                <a href="1" {% if records['page'] <=1 %} class="end" {% endif %}>&#8249;&#8249;</a>
-                <a href="{{ records['page'] - 1 }}" {% if records['page'] <=1 %} class="end" {% endif %}>&#8249;</a>
-                <span>{{ records["page"] }} / {{ records["pages"] }}</span>
-                <a href="{{ records['page'] + 1 }}" {% if records['page']>= records['pages'] %}
-                    class="end" {% endif
-                    %}>&#8250;</a>
-                <a href="{{ records['pages'] }}" {% if records['page']>= records['pages'] %}
-                    class="end" {% endif
-                    %}>&#8250;&#8250;</a>
+                <a href="1" {% if users['page'] <=1 %} class="end" {% endif %}>&#8249;&#8249;</a>
+                <a href="{{ users['page'] - 1 }}" {% if users['page'] <=1 %} class="end" {% endif %}>&#8249;</a>
+                <span>{{ users["page"] }} / {{ users["pages"] }}</span>
+                <a href="{{ users['page'] + 1 }}" {% if users['page']>= users['pages'] %}
+                    class="end" {% endif %}>&#8250;</a>
+                <a href="{{ users['pages'] }}" {% if users['page']>= users['pages'] %}
+                    class="end" {% endif %}>&#8250;&#8250;</a>
             </div>
         </div>
     </div>
