+++ templates/user_dashboard.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="{{ url_for('static', filename='dashboard.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='nav.css') }}">
     <link rel="icon" href="{{ url_for('static', filename='HSNU.png') }}">
-    <title>Dashboard</title>
+    <title>報修紀錄</title>
     {% if records["pages"] != 0 %}
         {% if records["page"] < 1 %}
         <script>window.location.replace("1");</script>
@@ -32,7 +32,7 @@
             <span class="right"><a href="/logout">登出</a></span>
             <span class="right"><a href="/user_setting">使用者設定</a></span>
         </nav>
-        <h1>Dashboard</h1>
+        <h1>報修紀錄</h1>
         <table>
             <thead>
                 <tr>
