+++ templates/user_setting.html
@@ -26,6 +26,7 @@
         <span class="right"><a href="/logout">登出</a></span>
         <span class="right"><a href="/user_setting" class="active">使用者設定</a></span>
     </nav>
+    <h1>使用者設定</h1>
     <div id="messages">{{ display_msg(messages) }}</div>
     <div id="user_setting" style="min-height: 100%;">
         <form action="/user_setting" method="POST">
